pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
passlib = {extras = ["argon2"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
